+++ src/RetroDev.OpenUI/Components/Base/UIComponent.cs
@@ -3,8 +3,8 @@ using RetroDev.OpenUI.Components.Core.AutoArea;
 using RetroDev.OpenUI.Core.Graphics;
 using RetroDev.OpenUI.Core.Graphics.Coordinates;
 using RetroDev.OpenUI.Core.Windowing.Events;
-using RetroDev.OpenUI.UI.Properties;
-using RetroDev.OpenUI.UI.Properties.Exceptions;
+using RetroDev.OpenUI.Presentation.Properties;
+using RetroDev.OpenUI.Presentation.Properties.Exceptions;
 
 namespace RetroDev.OpenUI.Components.Base;
