+++ src/RetroDev.OpenUI/Components/Containers/TreeBox.cs
@@ -3,7 +3,7 @@ using RetroDev.OpenUI.Components.Base;
 using RetroDev.OpenUI.Components.Core.AutoArea;
 using RetroDev.OpenUI.Components.Simple;
 using RetroDev.OpenUI.Core.Graphics.Coordinates;
-using RetroDev.OpenUI.UI.Properties;
+using RetroDev.OpenUI.Presentation.Properties;
 
 namespace RetroDev.OpenUI.Components.Containers;
