+++ src/RetroDev.OpenUI/Components/Shapes/Text.cs
@@ -2,8 +2,8 @@
 using RetroDev.OpenUI.Core.Graphics;
 using RetroDev.OpenUI.Core.Graphics.Coordinates;
 using RetroDev.OpenUI.Core.Windowing.Events;
-using RetroDev.OpenUI.UI;
-using RetroDev.OpenUI.UI.Properties;
+using RetroDev.OpenUI.Presentation;
+using RetroDev.OpenUI.Presentation.Properties;
 
 namespace RetroDev.OpenUI.Components.Shapes;
