+++ src/RetroDev.OpenUI/Presentation/Properties/IBinder{TValue}.cs
@@ -1,4 +1,4 @@
-﻿namespace RetroDev.OpenUI.UI.Properties;
+﻿namespace RetroDev.OpenUI.Presentation.Properties;
 
 // TODO: other than property binders we can have other types of binders?
 // e.g. DB binder where NotifySourceChanged() is storing to database
