+++ src/RetroDev.OpenUI/Presentation/Resources/EmbeddedWindowResources.cs
@@ -1,4 +1,4 @@
-﻿namespace RetroDev.OpenUI.UI.Resources;
+﻿namespace RetroDev.OpenUI.Presentation.Resources;
 
 /// <summary>
 /// Allows to locate a window resource.
