+++ src/RetroDev.OpenUI/UIDefinition/EAMLBinder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using System.Text.RegularExpressions;
 using RetroDev.OpenUI.Components.Base;
-using RetroDev.OpenUI.UI.Properties;
+using RetroDev.OpenUI.Presentation.Properties;
 using RetroDev.OpenUI.UIDefinition.Exceptions;
 
 // TODO: bindings
